+++ src/hotfolder.ts
@@ -5,10 +5,13 @@ import * as path from "path";
 import * as fileType from "file-type";
 import { helper } from "./helper";
 import { hotfolderSettings, settings } from "./settings";
+import { I18n } from "i18n";
 
 const fs = require("fs-extra");
 
+let i18n: any;
 let watchers = [];
+let dialogBox = null;
 
 export namespace hotfolder {
   export async function register() {
@@ -30,6 +33,23 @@ export namespace hotfolder {
       } catch (e) {}
 
       if (hotfolderPath.trim() != "") {
+        const importNotebook = await joplin.settings.value(
+          "importNotebook" + (hotfolderNr == 0 ? "" : hotfolderNr)
+        );
+
+        if (
+          importNotebook.trim() !== "" &&
+          (await helper.checkNotebookExist(importNotebook.trim())) === false
+        ) {
+          await hotfolder.showMsg(
+            i18n.__(
+              "error.notebookNotExist",
+              importNotebook.trim(),
+              hotfolderNr == 0 ? "" : hotfolderNr + 1
+            )
+          );
+        }
+
         let hotfolderWatcher = chokidar
           .watch(hotfolderPath, {
             persistent: true,
@@ -52,6 +72,40 @@ export namespace hotfolder {
     }
   }
 
+  export async function createDialogBox() {
+    dialogBox = await joplin.views.dialogs.create("hotfolderDialog");
+  }
+
+  export async function showMsg(msg: string, title: string = null) {
+    const html = [];
+
+    html.push(`<h3>Hotfolder plugin</h3>`);
+    if (title) {
+      html.push(`<p>${title}</p>`);
+    }
+    html.push(`<div id="msg">${msg}`);
+    html.push("</div>");
+    await joplin.views.dialogs.setButtons(dialogBox, [{ id: "ok" }]);
+    await joplin.views.dialogs.setHtml(dialogBox, html.join("\n"));
+    await joplin.views.dialogs.open(dialogBox);
+  }
+
+  export async function confLocale() {
+    const installationDir = await joplin.plugins.installationDir();
+    const localesDir = path.join(installationDir, "locales");
+    const joplinLocale = await joplin.settings.globalValue("locale");
+    i18n = new I18n({
+      locales: ["en_US", "de_DE"],
+      defaultLocale: "en_US",
+      fallbacks: { "en_*": "en_US" },
+      updateFiles: false,
+      retryInDefaultLocale: true,
+      syncFiles: true,
+      directory: localesDir,
+    });
+    i18n.setLocale(joplinLocale);
+  }
+
   export async function processFile(file: string, hotfolderNr: string) {
     const hotfolderSettings: hotfolderSettings = await settings.getHotfolder(
       hotfolderNr
@@ -70,9 +124,25 @@ export namespace hotfolder {
       let newBody = null;
       let noteTitle = fileName.replace(fileExt, "");
 
+      console.log(hotfolderSettings.notebookId);
+
+      if (
+        hotfolderSettings.importNotebook.trim() !== "" &&
+        (await helper.checkNotebookExist(
+          hotfolderSettings.importNotebook.trim()
+        )) === false
+      ) {
+        hotfolder.showMsg(
+          i18n.__(
+            "error.notebookNotExist",
+            hotfolderSettings.importNotebook.trim(),
+            parseInt(hotfolderNr) == 0 ? "" : parseInt(hotfolderNr) + 1
+          )
+        );
+      }
+
       if (
         hotfolderSettings.extensionsAddAsText
-          .toLowerCase()
           .split(/\s*,\s*/)
           .indexOf(fileExt) !== -1
       ) {
@@ -80,7 +150,8 @@ export namespace hotfolder {
         newNote = await hotfolder.importAsText(
           file,
           noteTitle,
-          hotfolderSettings.notebookId
+          hotfolderSettings.notebookId,
+          hotfolderSettings.textAsTodo
         );
       } else {
         console.info("Import as attachment");
@@ -107,7 +178,8 @@ export namespace hotfolder {
   export async function importAsText(
     file: string,
     noteTitle: string,
-    folder: string
+    folder: string,
+    todo: boolean
   ): Promise<any> {
     let fileBuffer = null;
     try {
@@ -121,6 +193,7 @@ export namespace hotfolder {
       body: fileBuffer.toString(),
       title: noteTitle,
       parent_id: folder,
+      is_todo: todo ? 1 : 0,
     });
   }
 
@@ -165,3 +238,5 @@ export namespace hotfolder {
     }
   }
 }
+
+export { i18n };