+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## not released
 
+## v1.1.0 (2023-06-10)
+
+- Fix: #24 mixed case file extentions in `Add as text` setting
+- Add: DirectoryPath selector for hotfolder path on Joplin >= v2.10.4
+- Add: Translation to German
+- Add: Option `Add text notes as Todo` to import text files as ToDo
+- Add: #29 Added warning message when importing to a non-existing notebook
+
 ## v1.0.1 (2021-08-12)
 
 - Fix: #13 `Hotfolder Path` is missing in settings and `Ignore Files` shows wrong value